presenter: split out the concerns
